+++ cypress/integration/Pokestats.spec.ts
@@ -43,13 +43,23 @@ describe('PokéStats', () => {
     cy.get('[data-testid="cause of fainting"]')
       .type('Died to crit')
       .should('have.value', 'Died to crit');
-    cy.get('[data-testid=move-1-4] > .search').type('Pound');
-    cy.get('[data-testid=move-2-4] > .search').type('Pound');
-    cy.get('[data-testid=move-3-4] > .search').type('Pound');
-    cy.get('[data-testid=move-4-4] > .search').type('Pound');
     cy.get('[data-testid=gender-4]').click();
     cy.get('[data-testid=gender-4] > .visible > :nth-child(2)').click();
     cy.get('[data-testid=nature-4]').type('Bold');
+    cy.get('[data-testid=move-summary]').click();
+    cy.get('[data-testid=move-1] > div').click();
+    cy.get('[data-testid=move-Pound]').click();
+    cy.get('[data-testid=move-2] > div').click();
+    cy.get('[data-testid="move-Karate Chop"]').click({ force: true });
+    cy.get('[data-testid=move-3] > div').click();
+    cy.get('[data-testid=filter-button]').click();
+    cy.get('[data-testid=filter-gen-1]').click();
+    cy.get('[data-testid=filter-type-WATER]').click();
+    cy.get('[data-testid=filter-button]').click();
+    cy.get('[data-testid=move-Surf]').click({ force: true });
+    cy.get('[data-testid=move-4] > div').click();
+    cy.get('[data-testid=move-Pound]').click();
+
     cy.contains('Save').click();
 
     cy.get('[data-testid=pokemon-5] > .search').type('Roselia');
@@ -58,7 +68,7 @@ describe('PokéStats', () => {
     cy.get('[data-testid=status-5]').click();
     cy.get('[data-testid=status-5] > .visible > :nth-child(5)').click();
 
-    cy.contains('PokéStats').click();
+    cy.contains('Stats').click();
     cy.contains('16%').should('exist');
     cy.get('.secondary > :nth-child(3) > .ui').click().should('have.text', '4');
     cy.get('.secondary > :nth-child(4) > .ui').should('have.text', '1');