+++ src/components/Type/Type.tsx
@@ -1,25 +1,28 @@
-import React from 'react';
+import React, { useCallback } from 'react';
+import useStore from 'store';
 import { TYPE_COLOR } from 'constants/colors';
-import { TPokemon } from 'constants/types';
+import TYPE_ICON from 'constants/icons';
+import { Type as TType } from 'constants/types';
 import styles from './Type.module.scss';
 
 interface TypeProps {
-  pokemon: TPokemon;
+  hideName?: boolean;
+  type: TType;
 }
 
-const Type: React.FC<TypeProps> = ({ pokemon }) => {
+const Type: React.FC<TypeProps> = ({ hideName = false, type }) => {
+  const showTypeModal = useStore(useCallback((state) => state.showTypeModal, []));
+
   return (
-    <div className={styles.container}>
-      <span>Type:</span>
-      <div className={styles.types}>
-        <span style={{ backgroundColor: TYPE_COLOR[pokemon?.type] }}>{pokemon?.type}</span>
-        {!!pokemon?.dualtype && (
-          <span style={{ backgroundColor: TYPE_COLOR[pokemon?.dualtype] }}>
-            {pokemon?.dualtype}
-          </span>
-        )}
-      </div>
-    </div>
+    <button
+      className={styles.type}
+      onClick={() => showTypeModal(type)}
+      style={{ backgroundColor: TYPE_COLOR[type] }}
+      type="button"
+    >
+      {TYPE_ICON[type]}
+      {!hideName && <div className={styles.text}>{type}</div>}
+    </button>
   );
 };